requirements.txt
BeautifulSoup4
requests
Pillow